upd files and save

+++ app/Http/Controllers/StaffController.php
@@ -318,24 +318,45 @@ class StaffController extends Controller
 
     public function uploadFile(Request $request)
     {
-        $staff_file = Staff_file::find($request->fileId);
 
-        Storage::disk('public')->delete($staff_file->file_path);
+        if ($request->fileId == 0) {
 
-        if ($request->hasFile('file')) {
-            $originalName = $request->file->getClientOriginalName();
-            $chunksName = explode("_", $originalName);
+            if ($request->hasFile('file')) {
+                $originalName = $request->file->getClientOriginalName();
+                $chunksName = explode("_", $originalName);
 
-            $fileName = time() . '_' . $request->fileTypeKey . '_' . $originalName;
-            $filePath = $request->file->storeAs('files', $fileName, 'public');
+                $fileName = time() . '_' . $request->fileTypeKey . '_' . $originalName;
+                $filePath = $request->file->storeAs('files', $fileName, 'public');
 
-            $label = 'default'; // Default or extract from filename
+                $label = 'default'; // Default or extract from filename
 
-            $staff_file->update([
-                'file_type' => $request->fileTypeKey, // You need to determine this differently
-                'file_path' => $filePath,
-                'expiration_date' => $request->expirationDate
-            ]);
+                $staff_file = Staff_file::create([
+                    'staff_id' => $request->staffId,
+                    'file_type' => $request->fileTypeKey, // You need to determine this differently
+                    'file_path' => $filePath,
+                    'expiration_date' => $request->expirationDate
+                ]);
+            }
+        } else {
+            $staff_file = Staff_file::find($request->fileId);
+
+            Storage::disk('public')->delete($staff_file->file_path);
+
+            if ($request->hasFile('file')) {
+                $originalName = $request->file->getClientOriginalName();
+                $chunksName = explode("_", $originalName);
+
+                $fileName = time() . '_' . $request->fileTypeKey . '_' . $originalName;
+                $filePath = $request->file->storeAs('files', $fileName, 'public');
+
+                $label = 'default'; // Default or extract from filename
+
+                $staff_file->update([
+                    'file_type' => $request->fileTypeKey, // You need to determine this differently
+                    'file_path' => $filePath,
+                    'expiration_date' => $request->expirationDate
+                ]);
+            }
         }
     }
 }
